+++ ferrodec-decimal64/src/lib.rs
@@ -25,9 +25,13 @@
 mod bid;
 mod classify;
 mod classify_types;
+mod convert;
 mod decimal;
+mod ops;
 mod status;
 
 pub use classify_types::IeeeClass;
+#[cfg(feature = "fmt")]
+pub use convert::{Engineering, ParseDecimalError};
 pub use decimal::{Decimal64, Decimal64BuildError};
 pub use status::{RoundingMode, Status};